requirements.txt
Flask>=3.0.0,<4.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
reportlab>=4.0.0,<5.0.0
supabase>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
